# plan, engine: fix push down of binary expressions in distributed optimizer